requirements.txt
airflow-clickhouse-plugin
alembic
annotated-types
anyio
apache-airflow
apache-airflow-providers-common-sql
apache-airflow-providers-ftp
apache-airflow-providers-http
apache-airflow-providers-imap
apache-airflow-providers-sqlite
apispec
argcomplete
asgiref
async-timeout
attrs
Babel
backoff
blinker
cachelib
cattrs
certifi
cffi
charset-normalizer
click
clickclick
clickhouse-driver
colorama
colorlog
ConfigUpdater
confluent-kafka
connexion
cron-descriptor
croniter
cryptography
Deprecated
dill
dnspython
docutils
email-validator
exceptiongroup
Flask
Flask-AppBuilder
Flask-Babel
Flask-Caching
Flask-JWT-Extended
Flask-Limiter
Flask-Login
Flask-Session
Flask-SQLAlchemy
Flask-WTF
frozenlist
google-re2
googleapis-common-protos
graphviz
grpcio
gunicorn
h11
httpcore
httpx
idna
importlib-metadata
importlib-resources
infi.clickhouse-orm
inflection
iso8601
itsdangerous
Jinja2
jsonschema
jsonschema-specifications
lazy-object-proxy
limits
linkify-it-py
lockfile
Mako
Markdown
markdown-it-py
MarkupSafe
marshmallow
marshmallow-oneofschema
marshmallow-sqlalchemy
mdit-py-plugins
mdurl
multidict
numpy
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
ordered-set
packaging
pandas
pathspec
pendulum
pluggy
prison
protobuf
psutil
pycparser
pydantic
pydantic_core
Pygments
PyJWT
python-daemon
python-dateutil
python-nvd3
python-slugify
pytz
pytzdata
PyYAML
referencing
requests
requests-toolbelt
rfc3339-validator
rich
rich-argparse
setproctitle
six
sniffio
SQLAlchemy
sqlalchemy-clickhouse
SQLAlchemy-JSONField
SQLAlchemy-Utils
sqlparse
tabulate
tenacity
termcolor
text-unidecode
typing_extensions
tzdata
tzlocal
uc-micro-py
unicodecsv
urllib3
Werkzeug
wrapt
WTForms
yarl
zipp
